requirements.txt
asgiref==3.8.1
autopep8==2.3.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==73.0.1
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
